[#164829365] Refactor tests to minimize repetition of called variables and isolated setup in a commons.py file
