+++ authors/apps/articles/tests/test_reviews.py
@@ -1,12 +1,6 @@
-from django.test import TestCase
-from django.urls import reverse
-from rest_framework import test, status
-from rest_framework.test import APIClient
-from authors.apps.articles.models import Article, ReviewsModel
-from authors.apps.authentication.models import User
+from .commons import *
 
-
-class TestReviewModel(TestCase):
+class TestReviewModel(BaseSetup):
     def setUp(self):
         user = User.objects.create(username="johndoe")
         article = Article.objects.create(
@@ -32,125 +26,19 @@ class TestReviewModel(TestCase):
         self.assertNotEqual(initial, updated)
 
 
-class ReviewTestCase(TestCase):
-    def setUp(self):
-        """
-        This setup creates and logs in a new user. After login the user
-        receives an auth token that is stored and will be used to access the
-        protected views contained in this test series. Lastly, two mock profile
-        sets for a user are made: the first being valid,
-        the second being invalid.
-        """
-        self.client = APIClient()
-        self.client2 = APIClient()
-        self.user = self.client.post(
-            reverse('authentication:user-signup'),
-            data={
-                "user": {
-                    "email": "demo@mail.com",
-                    "username": "Bob",
-                    "password": "Bob12345"
-                }
-            },
-            format="json"
-        )
-        self.user2 = self.client2.post(
-            reverse('authentication:user-signup'),
-            data={
-                "user": {
-                    "email": "pete@mail.com",
-                    "username": "Pete",
-                    "password": "Pete12345"
-                }
-            },
-            format="json"
-        )
-        test_user = User.objects.get(username="Bob")
-        test_user.is_verified = True
-        test_user.save()
-        test_user2 = User.objects.get(username="Pete")
-        test_user2.is_verified = True
-        test_user2.save()
-        self.login = self.client.post(
-            reverse('authentication:user-login'),
-            data={
-                "user": {
-                    "email": "demo@mail.com",
-                    "password": "Bob12345"
-                }
-            },
-            format="json"
-        )
-        self.login2 = self.client2.post(
-            reverse('authentication:user-login'),
-            data={
-                "user": {
-                    "email": "pete@mail.com",
-                    "password": "Pete12345"
-                }
-            },
-            format="json"
-        )
-        self.token = self.login.data['token']
-        self.token2 = self.login2.data['token']
-        self.client.credentials(HTTP_AUTHORIZATION='Bearer ' + self.token)
-        self.client2.credentials(HTTP_AUTHORIZATION='Bearer ' + self.token2)
+class ReviewTestCase(BaseSetup):
 
     def test_user_can_create_review(self):
-        response = self.client.post(
-            reverse('articles:create-list'),
-            data={
-                "article": {
-                    "title": "Test title",
-                    "body": "This is a very awesome article on testing tests",
-                    "description": "Written by testing tester",
-                }
-            },
-            format="json"
-        )
-
+        response = self.post_article()
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        response = self.client2.post(
-            reverse('articles:review', kwargs={
-                "slug": Article.objects.get().slug
-            }),
-            data={
-                "review": {
-                    "review_body": "I really liked the article",
-                    "rating_value": 4
-                }
-            },
-            format="json"
-        )
+        response = self.post_review()
         print(response.data)
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
 
     def test_user_can_get_reviews(self):
-        response = self.client.post(
-            reverse('articles:create-list'),
-            data={
-                "article": {
-                    "title": "Test title",
-                    "body": "This is a very awesome article on testing tests",
-                    "description": "Written by testing tester",
-                }
-            },
-            format="json"
-        )
-
+        response = self.post_article()
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        response = self.client2.post(
-            reverse('articles:review', kwargs={
-                "slug": Article.objects.get().slug
-            }),
-            data={
-                "review": {
-                    "review_body": "I really liked the article",
-                    "rating_value": 4
-                }
-            },
-            format="json"
-        )
+        response = self.post_review()
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
         response = self.client2.get(
             reverse('articles:review', kwargs={
@@ -162,31 +50,10 @@ class ReviewTestCase(TestCase):
         self.assertContains(response, "I really liked the article")
 
     def test_user_can_edit_review(self):
-        response = self.client.post(
-            reverse('articles:create-list'),
-            data={
-                "article": {
-                    "title": "Test title",
-                    "body": "This is a very awesome article on testing tests",
-                    "description": "Written by testing tester",
-                }
-            },
-            format="json"
-        )
+        response = self.post_article()
 
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        response = self.client2.post(
-            reverse('articles:review', kwargs={
-                "slug": Article.objects.get().slug
-            }),
-            data={
-                "review": {
-                    "review_body": "I really liked the article",
-                    "rating_value": 4
-                }
-            },
-            format="json"
-        )
+        response = self.post_review()
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
         response = self.client2.put(reverse('articles:review', kwargs={
             "slug": Article.objects.get().slug
@@ -200,31 +67,10 @@ class ReviewTestCase(TestCase):
         self.assertEqual(response.status_code, status.HTTP_200_OK)
 
     def test_user_can_delete_review(self):
-        response = self.client.post(
-            reverse('articles:create-list'),
-            data={
-                "article": {
-                    "title": "Test title",
-                    "body": "This is a very awesome article on testing tests",
-                    "description": "Written by testing tester",
-                }
-            },
-            format="json"
-        )
+        response = self.post_article()
 
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        response = self.client2.post(
-            reverse('articles:review', kwargs={
-                "slug": Article.objects.get().slug
-            }),
-            data={
-                "review": {
-                    "review_body": "I really liked the article",
-                    "rating_value": 4
-                }
-            },
-            format="json"
-        )
+        response = self.post_review()
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
         response = self.client2.delete(reverse('articles:review', kwargs={
             "slug": Article.objects.get().slug
@@ -236,51 +82,20 @@ class ReviewTestCase(TestCase):
             }),
             format="json"
         )
-        print(response_delete.data)
         self.assertEqual(response_delete.status_code,
                          status.HTTP_404_NOT_FOUND)
         self.assertRaises(Exception)
 
     def test_unautheticated_user_cant_create_review(self):
-        response = self.client.post(
-            reverse('articles:create-list'),
-            data={
-                "article": {
-                    "title": "Test title",
-                    "body": "This is a very awesome article on testing tests",
-                    "description": "Written by testing tester",
-                }
-            },
-            format="json"
-        )
+        response = self.post_article()
 
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
         self.client2.logout()
-        response = self.client2.post(
-            reverse('articles:review', kwargs={
-                "slug": Article.objects.get().slug
-            }),
-            data={
-                "review": {
-                    "review_body": "I really liked the article",
-                    "rating_value": 4
-                }
-            },
-            format="json"
-        )
+        response = self.post_review()
         self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+
     def test_review_creation_with_wrong_values(self):
-        response = self.client.post(
-            reverse('articles:create-list'),
-            data={
-                "article": {
-                    "title": "Test title",
-                    "body": "This is a very awesome article on testing tests",
-                    "description": "Written by testing tester",
-                }
-            },
-            format="json"
-        )
+        response = self.post_article()
 
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
         response = self.client2.post(
@@ -295,22 +110,10 @@ class ReviewTestCase(TestCase):
             },
             format="json"
         )
-        print(response.data)
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
 
     def test_user_cannot_review_own_article(self):
-        response = self.client.post(
-            reverse('articles:create-list'),
-            data={
-                "article": {
-                    "title": "Test title",
-                    "body": "This is a very awesome article on testing tests",
-                    "description": "Written by testing tester",
-                }
-            },
-            format="json"
-        )
-
+        response = self.post_article()
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
         response = self.client.post(
             reverse('articles:review', kwargs={
@@ -319,10 +122,9 @@ class ReviewTestCase(TestCase):
             data={
                 "review": {
                     "review_body": "I really liked the article",
-                    "rating_value": 4
+                    "rating_value": 5
                 }
             },
             format="json"
         )
-        print(response.data)
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
